Updated board type
because there where some changes to the compatibility of the ESP32 single core chips, i change the config accordingly so that it should work again. Also i added the NTC Sensor for overheating protection, but could not figure out the correct resistor values, which is why I’ve mor or less just guest them. The onboard button (rest button) i also added to the config file